# server/qa-baselines/openapi.current.json
{
  "components": {
    "schemas": {
      "ErrorEnvelope": {
        "additionalProperties": false,
        "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}.",
        "properties": {
          "error": {
            "additionalProperties": false,
            "properties": {
              "code": {
                "description": "Stable code from the ERROR_CODES catalog.",
                "type": "string"
              },
              "details": {
                "additionalProperties": true,
                "type": "object"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "novel_engine_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Self-hosted writing studio API (TypeScript rewrite).",
    "title": "Novel Engine API",
    "version": "0.6.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/projects": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "projects": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "created_at": {
                            "type": "string"
                          },
                          "description": {
                            "type": "string"
                          },
                          "documents": {
                            "items": {
                              "additionalProperties": false,
                              "properties": {
                                "beat_ref": {
                                  "nullable": true,
                                  "type": "string"
                                },
                                "content_markdown": {
                                  "type": "string"
                                },
                                "created_at": {
                                  "type": "string"
                                },
                                "current_revision_id": {
                                  "type": "string"
                                },
                                "id": {
                                  "type": "string"
                                },
                                "kind": {
                                  "enum": [
                                    "chapter",
                                    "outline",
                                    "character",
                                    "world",
                                    "note"
                                  ],
                                  "type": "string"
                                },
                                "lore_status": {
                                  "enum": [
                                    "draft",
                                    "stable",
                                    "deprecated"
                                  ],
                                  "nullable": true,
                                  "type": "string"
                                },
                                "metadata": {
                                  "additionalProperties": true,
                                  "type": "object"
                                },
                                "position": {
                                  "type": "integer"
                                },
                                "project_id": {
                                  "type": "string"
                                },
                                "revision_source": {
                                  "enum": [
                                    "author",
                                    "ai-accepted",
                                    "restore"
                                  ],
                                  "type": "string"
                                },
                                "title": {
                                  "type": "string"
                                },
                                "updated_at": {
                                  "type": "string"
                                },
                                "volume_id": {
                                  "nullable": true,
                                  "type": "string"
                                },
                                "word_count": {
                                  "type": "integer"
                                }
                              },
                              "required": [
                                "id",
                                "project_id",
                                "kind",
                                "title",
                                "position",
                                "volume_id",
                                "beat_ref",
                                "lore_status",
                                "current_revision_id",
                                "content_markdown",
                                "metadata",
                                "revision_source",
                                "word_count",
                                "created_at",
                                "updated_at"
                              ],
                              "type": "object"
                            },
                            "type": "array"
                          },
                          "id": {
                            "type": "string"
                          },
                          "import_hash": {
                            "nullable": true,
                            "type": "string"
                          },
                          "settings": {
                            "additionalProperties": true,
                            "type": "object"
                          },
                          "title": {
                            "type": "string"
                          },
                          "updated_at": {
                            "type": "string"
                          },
                          "volumes": {
                            "items": {
                              "additionalProperties": false,
                              "properties": {
                                "created_at": {
                                  "type": "string"
                                },
                                "id": {
                                  "type": "string"
                                },
                                "position": {
                                  "type": "integer"
                                },
                                "project_id": {
                                  "type": "string"
                                },
                                "title": {
                                  "type": "string"
                                },
                                "updated_at": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "id",
                                "project_id",
                                "title",
                                "position",
                                "created_at",
                                "updated_at"
                              ],
                              "type": "object"
                            },
                            "type": "array"
                          }
                        },
                        "required": [
                          "id",
                          "title",
                          "description",
                          "settings",
                          "import_hash",
                          "created_at",
                          "updated_at"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "projects"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          }
        }
      },
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "description": {
                    "default": "",
                    "maxLength": 10000,
                    "type": "string"
                  },
                  "title": {
                    "maxLength": 240,
                    "minLength": 1,
                    "type": "string"
                  }
                },
                "required": [
                  "title"
                ],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "created_at": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "documents": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "beat_ref": {
                            "nullable": true,
                            "type": "string"
                          },
                          "content_markdown": {
                            "type": "string"
                          },
                          "created_at": {
                            "type": "string"
                          },
                          "current_revision_id": {
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          },
                          "kind": {
                            "enum": [
                              "chapter",
                              "outline",
                              "character",
                              "world",
                              "note"
                            ],
                            "type": "string"
                          },
                          "lore_status": {
                            "enum": [
                              "draft",
                              "stable",
                              "deprecated"
                            ],
                            "nullable": true,
                            "type": "string"
                          },
                          "metadata": {
                            "additionalProperties": true,
                            "type": "object"
                          },
                          "position": {
                            "type": "integer"
                          },
                          "project_id": {
                            "type": "string"
                          },
                          "revision_source": {
                            "enum": [
                              "author",
                              "ai-accepted",
                              "restore"
                            ],
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "updated_at": {
                            "type": "string"
                          },
                          "volume_id": {
                            "nullable": true,
                            "type": "string"
                          },
                          "word_count": {
                            "type": "integer"
                          }
                        },
                        "required": [
                          "id",
                          "project_id",
                          "kind",
                          "title",
                          "position",
                          "volume_id",
                          "beat_ref",
                          "lore_status",
                          "current_revision_id",
                          "content_markdown",
                          "metadata",
                          "revision_source",
                          "word_count",
                          "created_at",
                          "updated_at"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "id": {
                      "type": "string"
                    },
                    "import_hash": {
                      "nullable": true,
                      "type": "string"
                    },
                    "settings": {
                      "additionalProperties": true,
                      "type": "object"
                    },
                    "title": {
                      "type": "string"
                    },
                    "updated_at": {
                      "type": "string"
                    },
                    "volumes": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "created_at": {
                            "type": "string"
                          },
                          "id": {
                            "type": "string"
                          },
                          "position": {
                            "type": "integer"
                          },
                          "project_id": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          },
                          "updated_at": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "project_id",
                          "title",
                          "position",
                          "created_at",
                          "updated_at"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "id",
                    "title",
                    "description",
                    "settings",
                    "import_hash",
                    "created_at",
                    "updated_at",
                    "documents",
                    "volumes"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          }
        }
      }
    },
...
    "/api/projects/{projectId}/search": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "projectId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "results": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "document_id": {
                            "type": "string"
                          },
                          "excerpt": {
                            "type": "string"
                          },
                          "title": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "document_id",
                          "title",
                          "excerpt"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": [
                    "results"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Default Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unified error envelope: every API failure renders as {error:{code,message,details?}}."
          }
        }
      }
    },
...
  }
}